add procted db
