+++ test/distributed/cases/tenant/privilege/protected_database.result
@@ -1,0 +1,46 @@
+set global enable_privilege_cache = off;
+drop database if exists protected_bvt_sys;
+set global protected_databases = 'protected_bvt_sys';
+create database protected_bvt_sys;
+create table protected_bvt_sys.t1(a int);
+drop table protected_bvt_sys.t1;
+drop database protected_bvt_sys;
+set global protected_databases = 'protected_bvt_unused';
+drop account if exists protected_bvt_acc;
+create account protected_bvt_acc ADMIN_NAME 'admin' IDENTIFIED BY '111';
+set global protected_databases = 'protected_bvt_db,protected_bvt_new,CamelDB';
+set global protected_databases = '';
+internal error: protected_databases cannot be cleared directly
+create database protected_bvt_db;
+create table protected_bvt_db.t1(a int);
+insert into protected_bvt_db.t1 values(1);
+create role protected_bvt_writer;
+grant create database,drop database on account * to protected_bvt_writer;
+grant all on database protected_bvt_db to protected_bvt_writer;
+grant all on table protected_bvt_db.* to protected_bvt_writer;
+create user protected_bvt_user identified by '111' default role protected_bvt_writer;
+select * from protected_bvt_db.t1;
+a
+1
+create database protected_bvt_normal;
+drop database protected_bvt_normal;
+create database protected_bvt_new;
+internal error: do not have privilege to execute the statement
+create table protected_bvt_db.t2(a int);
+internal error: do not have privilege to execute the statement
+insert into protected_bvt_db.t1 values(2);
+internal error: do not have privilege to execute the statement
+truncate table protected_bvt_db.t1;
+internal error: do not have privilege to execute the statement
+drop table protected_bvt_db.t1;
+internal error: do not have privilege to execute the statement
+drop database protected_bvt_db;
+internal error: do not have privilege to execute the statement
+create database `cameldb`;
+drop database `cameldb`;
+drop user if exists protected_bvt_user;
+drop role if exists protected_bvt_writer;
+drop database protected_bvt_db;
+set global protected_databases = 'protected_bvt_unused';
+drop account protected_bvt_acc;
+set global enable_privilege_cache = on;
